Detach Groups to GroupsMixin, filter->filter_string
