+++ gvm/protocols/gmpv208/__init__.py
@@ -70,6 +70,7 @@ from gvm.protocols.gmpv208.entities.filter import (
     FilterType,
     get_filter_type_from_string,
 )
+from gvm.protocols.gmpv208.entities.groups import GroupsMixin
 from gvm.protocols.gmpv208.entities.hosts import (
     HostsMixin,
     HostsOrdering,
@@ -146,6 +147,7 @@ class Gmp(
     CredentialsMixin,
     FeedsMixin,
     FiltersMixin,
+    GroupsMixin,
     HostsMixin,
     NotesMixin,
     OperatingSystemsMixin,
