+++ tests/protocols/gmpv208/testcmds/test_get_groups.py
@@ -23,8 +23,8 @@ class GmpGetGroupsTestCase:
 
         self.connection.send.has_been_called_with('<get_groups/>')
 
-    def test_get_groups_with_filter(self):
-        self.gmp.get_groups(filter='foo=bar')
+    def test_get_groups_with_filter_string(self):
+        self.gmp.get_groups(filter_string='foo=bar')
 
         self.connection.send.has_been_called_with(
             '<get_groups filter="foo=bar"/>'
